requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
line-bot-sdk>=3.5.0
pydantic-settings>=2.0.0
httpx>=0.25.0
pytz>=2024.1
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0
Pillow>=10.0.0
imagehash>=4.3.1
